Implement efficient LRU cache
An LRU cache is a key-value map that stores data in a way that makes it easy to remove the least recently used item. It uses a doubly linked list to keep track of the order in which items were last used.

When a new item is added to the cache, a list node is created and added to both the map and the doubly linked list. When an item is retrieved from the cache, its corresponding list node is moved to the front of the list, which marks it as the most recently used item.

If the cache runs out of space, the least recently used item (which is at the end of the list) is removed to make room for the new item.

The Add and Get functions of the LRU cache have a runtime complexity of O(1), meaning that they execute in constant time regardless of the size of the cache. The space complexity is O(N), where N is the maximum number of items that the cache can hold.

Signed-off-by: Arunprasad Rajkumar <ar.arunprasad@gmail.com>